move logm to SPD + add classmethod decorators
