+++ tests/test_symmetric_matrices.py
@@ -59,17 +59,6 @@ class TestSymmetricMatricesMethods(geomstats.tests.TestCase):
                               [0., 0., e]]])
         self.assertAllClose(result, expected)
 
-    def test_logm(self):
-        """Test of logm method."""
-        sym_n = SymmetricMatrices(self.n)
-        expected = gs.array([[[0., 1., 0.], [1., 0., 0.], [0., 0., 1.]]])
-        c = math.cosh(1)
-        s = math.sinh(1)
-        e = math.exp(1)
-        v = gs.array([[[c, s, 0.], [s, c, 0.], [0., 0., e]]])
-        result = sym_n.logm(v)
-        self.assertAllClose(result, expected)
-
     def test_powerm(self):
         """Test of powerm method."""
         sym_n = SymmetricMatrices(self.n)
